Buffer Overflow causes by Pointers
In the renderBattle function from my program pointers are used to change data which may case can become a risk. If there is a scenario where the size exceeds the buffer size it may lead to the program crashing or worse any bad actor may execute code. While not committed to the code, a fix I am working on is by using vectors rather than the raw pointers.